kiss: Remove KISS_RM
I never really liked how this was implemented and it's
rather limited in its use. It was cumbersome to edit
the list to make even a minor change.

Instead, I now recommend using 'KISS_HOOK'. It gives
you total control. KISS_RM can be implemented on a
per-package basis, for all packages overall or
conditionally.

You're also not limited to a simple list of removals.
You can do whatever you like pre- and post- build.

Here's an example script to get you started. There
is more information in the manual and README.

case $TYPE in
    pre-build)
        case $PKG in
           zlib) export CFLAGS="-Os -static" ;;
           curl) export CFLAGS="-O3" ;;
        esac
    ;;

    post-build)
        : "${DEST:?DEST is unset}"

        rm -rf "$DEST/usr/share/doc"
        rm -rf "$DEST/usr/share/gettext"
    ;;
esac