kiss: simplify depends finder file handling.
Removes the use of multiple temporary files and the
number of times they're shuffled around. POSIX sort
is neat.

This should bring minor speedups to the dependency
finder on top of installing GNU grep. I'm sure this
can be further simplified.

I'll be tackling the reduction of checked files
next. Wish me luck.