kiss: Drop KISS_AUDIT for now.
I'm going to re-implement this feature in a
smarter way. It doesn't currently work for
partial updates nor does it work at all for
some users.

Some kind of persistent logging would be
far better as we'll be able to view a diff
regardless of the current pull's status.

The best way of doing this will be to grab
a diff of each package pending an update.
This way it's only information useful to
the user.

Fun stuff.