kiss: Configurable and dynamic tarball compression.
This allows you to swap between gzip and xz compression via
the new environment variable ('KISS_COMPRESS'). As of this
commit, new builds will use xz compression (making use of
all cores on the machine).

Other compression methods can easily be added by adding two
simple lines to the script. Your existing package cache will
continue to be used as the package manager will use whatever
tarball is available (for the package and version it is
looking for).