kiss: add $KISS_LIBRARY variable, to allow kiss to be sourced
Setting KISS_LIBRARY to non-zero will disable main, setting it to 2 will
disable die from exiting in certain cases, just returning 1 instead.
die() has been given a new first argument, which should be 0 when
exiting is to be disabled when KISS_LIBRARY is 2, and non-zero to force
exit regardless.
To maintain current behaviour until certain die cases are chosen to not
exit when KISS_LIBRARY is 2, all current uses of die have been given the
1 argument to force exit.
The value for KISS_LIBRARY to disable certain exits being specifically 2
and not just above 1 is deliberately to allow future values to be added
to invoke different behaviour.