kiss: Drop hardcoded musl reference.
This replaces the package based exemption with one based on C and
POSIX shared library names, removing the hardcoded reference to
musl and adding support for more C libraries.

This needs testing with non-musl systems. Specifically, the
dependency fixer which runs post-build in 'kiss build' should /not/
display anything related to your C library. If it does, please open
an issue.