kiss: drop filenames from checksums files
With this change, checksums generated by the package manager will no
longer contain the second field (the basename of the file). The second
field is quite useless as ordering is already fixed.

This change is backwards compatible and existing checksums files will
containue to function as normal. As checksums are regenerated (with
updates to packages), the new format will be used.

The compatibility works via checking /only the first column/ in each
checksums file. This is the only part of the file which must match
1:1 as ordering is always the same.

Notes about etcsums:

This change also affects the etcsums file which is used for the
3-way checksum verification for files in /etc/.

Existing etcsums will simply be invalidated with the safest path chosen
(.new file is created) until they are regenerated via a package build.

Ordering is now fixed inside etcsums files. This makes parsing them
faster and more reliable.

This also fixes a theoretical bug with naming conflicts during parsing
of etcsums.