kiss: add support for directories and absolute file paths as sources
This adds support for both absolute and relative directories in
sources files. Relative file sources have always been supported,
this just extends it to directories.

Absolute paths to directories and files in sources is totally new.

Directories will have their /contents/ copied to the initial build
directory. The second field in the sources file can be used to set
destination location.

Checksums will not be validated for directories of any type. Absolute
file paths do have their checksums validated.

Notes:

The package manager will check to see if the source is a directory
or not. There is no need to have a trailing forward slash (though it
might make sense to enforce this anyway, we'll see).