kiss: Add handling of /etc/ files to removal action.
If the user has modified the file in /etc/, it will not be removed
and a message will be printed to the screen. If the file is
unmodified, it will be removed like any other.

This uses the etcsums file for comparison of checksums. The file
is also used during installation handling of files in /etc/.