Faster conflict detection
Replace the use of grep to determine the unique dependencies with
sort/comm.

While testing this discovered that a bug when generating the list of
conflicts. If we had a list like "A B C", where B is the current
package, we'd search " A B C " for " B " and replace it with "" so we'd
end up with " AC " which is incorrect. That code has been modified to
replace with " " such that we'd keep A and C separate.