kiss: drop subshell usage from pkg_find
The subshell prevents die() from exiting on error which causes
issues with non-existent packages.

This modifies the function to store the path in $repo_dir which
is then used by callers. Had to shuffle some things around and
fix issues with the new behavior.

New behavior is that non-existent packages are caught ASAP.
Please let me know if there are any issues.