kiss-export: Remove utility
This utility must be rewritten (with added complexity). There is
no portable way to tell tar to add a directory but not include its
contents (unless it is empty).

The simple approach of giving tar the package's manifest is not
workable without omitting directories. This causes issues when
installing an exported tarball and target directories do not exist.

The most obvious solution is to copy all of the package's files
to a temporary directory and then pass that to tar. This would
work fine but slow the process down substantially.

Another solution will be sought and if not possible the above will
be implemented instead.

Closes #254