kiss-outdated: check repology in parallel
The biggest change from the original implementation is that chekcing for
updates is now done in parallel via xargs. This provides a massive speed
increase for even small repositories.

Old times:
core  -   17.141s
extra - 1m29.778s

New times:
core  -   1.689s
extra -   4.246s

Where 'core' and 'extra' are the core and extra repositories in the main
kisslinux repo.

One thing that is regressed in this patch is the maintainability of
kiss-outdated is now reduced as much of the script is now contained as a
string argument. Ideally this could be fixed by splitting up the script
to be two new scripts.