get owner of file from our C program
Replace parsing the output of "ls -ld" to a small program which just
prints the username of the owner of a file. Parsing ls(1) sometimes
caused truncation at 8 characters. The C file must be compiled when kiss
is installed and the binary is expected to be in /usr/lib/kiss. This
directory is prepended to $PATH and the new PATH is exported, so that
the tools can also be used from hooks.